eval/requirements.txt
numpy>=1.23.4
openai==1.3.5
requests
tqdm
pandas>=1.5.3
gradio==4.15.0
tiktoken
rich
portalocker
torch==2.1.2
torchvision==0.16.2
timeout-decorator
datasets
transformers==4.39.2
opencv-python>=4.4.0.46
typing_extensions==4.7.1
pillow
omegaconf
matplotlib
einops
sentencepiece
protobuf
sty
xtuner
huggingface_hub
visual_genome
pycocoevalcap
openpyxl
seaborn
tabulate
xlsxwriter